Update variable type to fix incompatibility
- Change the type of the 'one' variable from 'int' to 'u_long' to avoid type incompatibility
- Adjust the 'ioctlsocket' function call to use the correct type

+++ librabbitmq/amqp_socket.c
@@ -291,7 +291,7 @@ int amqp_open_socket_noblock(char const *hostname, int portnumber,
 
 #ifdef _WIN32
 static int connect_socket(struct addrinfo *addr, amqp_time_t deadline) {
-  int one = 1;
+  u_long one = 1;
   SOCKET sockfd;
   int last_error;
